Title: 최솟값 3개, Time: 65ms, Memory: 1MB, Status: Passed - CodeTree

+++ 241219/최솟값 3개/top-3-smallest-number.cpp
@@ -22,7 +22,7 @@ int main(){
             pq.pop();
             int c=pq.top();
             pq.pop();
-            cout<<a*b*c<<endl;
+            cout<<(long long)a*b*c<<endl;
             pq.push(a);
             pq.push(b);
             pq.push(c);
